requirements.txt
python-telegram-bot==13.7
tweepy==3.10.0
web3
zksync2
python-dateutil==2.8.2
pytz==2021.1
pyyaml
